StreamEmbedXXX into embeddings.py, EnsPredHead into engines
