+++ src/weathergen/model/engines.py
@@ -15,11 +15,12 @@ from weathergen.model.attention import (
     MultiSelfAttentionHead_Local,
     MultiSelfAttentionHead_Varlen,
 )
-from weathergen.model.layers import (
-    MLP,
+from weathergen.model.embeddings import (
     StreamEmbedLinear,
     StreamEmbedTransformer,
 )
+from weathergen.model.layers import MLP
+
 from weathergen.utils.config import Config
 
 
@@ -289,3 +290,47 @@ class ForecastingEngine:
                     )
                 )
         return self.fe_blocks
+
+
+
+
+class EnsPredictionHead(torch.nn.Module):
+    #########################################
+    def __init__(
+        self, dim_embed, dim_out, ens_num_layers, ens_size, norm_type="LayerNorm", hidden_factor=2
+    ):
+        """Constructor"""
+
+        super(EnsPredictionHead, self).__init__()
+
+        dim_internal = dim_embed * hidden_factor
+        # norm = torch.nn.LayerNorm if norm_type == "LayerNorm" else RMSNorm
+        enl = ens_num_layers
+
+        self.pred_heads = torch.nn.ModuleList()
+        for i in range(ens_size):
+            self.pred_heads.append(torch.nn.ModuleList())
+
+            # self.pred_heads[-1].append( norm( dim_embed))
+            self.pred_heads[-1].append(
+                torch.nn.Linear(dim_embed, dim_out if enl == 1 else dim_internal)
+            )
+
+            for i in range(ens_num_layers - 1):
+                self.pred_heads[-1].append(torch.nn.GELU())
+                self.pred_heads[-1].append(
+                    torch.nn.Linear(dim_internal, dim_out if enl - 2 == i else dim_internal)
+                )
+
+    #########################################
+    @torch.amp.custom_fwd(cast_inputs=torch.float32, device_type="cuda")
+    def forward(self, toks):
+        preds = []
+        for pred_head in self.pred_heads:
+            cpred = toks
+            for block in pred_head:
+                cpred = block(cpred)
+            preds.append(cpred)
+        preds = torch.stack(preds, 0)
+
+        return preds
